requirements.txt
gradio>=4.0.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
umap-learn>=0.5.4
scikit-learn>=1.3.0
datasets>=2.14.0
huggingface-hub>=0.17.0
tqdm>=4.66.0
python-dotenv>=1.0.0

